requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
anyio=3.5.0=py39hf3d152e_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py39h3811e60_1
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pyhd8ed1ab_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.10.0=pyha770c72_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
brotlipy=0.7.0=py39hb9d737c_1004
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pypi_0
certifi=2021.10.8=py39hf3d152e_2
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
cryptography=36.0.2=py39hd97740a_1
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd8ed1ab_0
debugpy=1.5.1=py39he80948d_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
flit-core=3.7.1=pyhd8ed1ab_0
fonttools=4.31.2=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
idna=3.3=pyhd8ed1ab_0
igl=2.2.1=py39h86c447d_1
importlib-metadata=4.11.3=py39hf3d152e_1
importlib_resources=5.6.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py39h06a4308_0
ipympl=0.9.1=pyhd8ed1ab_0
ipython=8.2.0=py39h06a4308_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
jedi=0.18.1=py39h06a4308_1
jinja2=3.1.1=pyhd8ed1ab_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter_client=7.2.1=pyhd8ed1ab_0
jupyter_core=4.9.2=py39hf3d152e_0
jupyter_server=1.16.0=pyhd8ed1ab_1
jupyterlab=3.3.2=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.12.0=pyhd8ed1ab_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
kiwisolver=1.4.2=py39hf939315_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_14
libgfortran-ng=11.2.0=h69a702a_14
libgfortran5=11.2.0=h5c6108e_14
libgomp=11.2.0=h1d223b6_14
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=he4da1e4_14
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
littleutils=0.2.2=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
markupsafe=2.1.1=py39hb9d737c_1
matplotlib=3.5.1=pypi_0
matplotlib-base=3.4.3=py39h2fa2bec_2
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h3811e60_1005
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=pypi_0
nbclassic=0.3.7=pyhd8ed1ab_0
nbclient=0.5.13=pyhd8ed1ab_0
nbconvert=6.4.5=pyhd8ed1ab_2
nbconvert-core=6.4.5=pyhd8ed1ab_2
nbconvert-pandoc=6.4.5=pyhd8ed1ab_2
nbformat=5.3.0=pyhd8ed1ab_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=2.7.1=pypi_0
notebook=6.4.10=pyha770c72_0
notebook-shim=0.1.0=pyhd8ed1ab_0
numpy=1.21.2=py39h20f2e39_0
numpy-base=1.21.2=py39h79a1101_0
oauthlib=3.2.0=pypi_0
ogb=1.3.3=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1o=h166bdaf_0
outdated=0.2.1=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.2=pypi_0
pandoc=2.17.1.1=ha770c72_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
prometheus_client=0.14.0=pyhd8ed1ab_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.7=pyhd8ed1ab_0
pyqt5=5.15.6=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.10.1=pypi_0
pyrsistent=0.18.1=py39hb9d737c_1
pysocks=1.7.1=py39hf3d152e_5
python=3.9.12=h12debd9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pyhd8ed1ab_0
pyzmq=22.3.0=py39headdf64_2
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=py39hee8e79c_1
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py39hf3d152e_3
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.2=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.13.3=py39hf3d152e_1
testpath=0.6.0=pyhd8ed1ab_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torch-geometric=2.0.4=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.13=pypi_0
torchaudio=0.11.0=py39_cu113
torchsummary=1.5.1=pypi_0
torchvision=0.12.0=py39_cu113
tornado=6.1=py39hb9d737c_3
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=pyhd8ed1ab_0
vtk=9.1.0=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py_1
websocket-client=1.3.2=pyhd8ed1ab_0
werkzeug=2.1.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=py39hf3d152e_0
wslink=1.5.1=pypi_0
xeus=2.4.1=h70bab47_0
xeus-python=0.13.7=py39h800c839_0
xeus-python-shell=0.2.0=pyhd8ed1ab_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
